Move LazyList and Map.mapValue rules to separate rules (fix #56)
