+++ scalafix/input/src/main/scala/fix/SetMapSrc.scala
@@ -1,5 +1,5 @@
 /*
-rule = "scala:fix.Scalacollectioncompat_newcollections"
+rule = "scala:fix.Stable"
  */
 package fix
 
@@ -8,5 +8,4 @@ class SetMapSrc(set: Set[Int], map: Map[Int, Int]) {
   map + (2 -> 3, 3 -> 4)
   (set + (2, 3)).toString
   set + (2, 3) - 4
-  map.mapValues(_ + 1)
 }
