+++ scalafix/input/src/main/scala/fix/UnstableMapValuesSrc.scala
@@ -1,0 +1,8 @@
+/*
+rule = "scala:fix.UnstableMapValues"
+ */
+package fix
+
+class UnstableMapValuesSrc(map: Map[Int, Int]) {
+  map.mapValues(_ + 1)
+}
