+++ scalafix/output/src/main/scala/fix/CopyToBufferSrc.scala
@@ -1,3 +1,6 @@
+
+
+
 package fix
 
 import scala.collection.mutable
