+++ scalafix/output/src/main/scala/fix/RetainSrc.scala
@@ -1,3 +1,6 @@
+
+
+
 package fix
 
 import scala.collection.mutable.{Map, Set}
