+++ scalafix/output/src/main/scala/fix/SetMapSrc.scala
@@ -8,5 +8,4 @@ class SetMapSrc(set: Set[Int], map: Map[Int, Int]) {
   map + (2 -> 3) + (3 -> 4)
   (set + 2 + 3).toString
   set + 2 + 3 - 4
-  map.mapValues(_ + 1).toMap
 }
