+++ scalafix/output/src/main/scala/fix/TraversableSrc.scala
@@ -1,3 +1,6 @@
+
+
+
 package fix
 
 object TraversableSrc {
