+++ scalafix/output/src/main/scala/fix/UnstableMapValuesSrc.scala
@@ -1,0 +1,8 @@
+
+
+
+package fix
+
+class UnstableMapValuesSrc(map: Map[Int, Int]) {
+  map.mapValues(_ + 1).toMap
+}
