+++ scalafix/output/src/main/scala/fix/UnstableStreamToLazyListSrc.scala
@@ -1,6 +1,9 @@
+
+
+
 package fix
 
-object StreamSrc {
+class UnstableStreamToLazyListSrc() {
   val s = LazyList(1, 2, 3)
   s.lazyAppendedAll(List(4, 5, 6))
   1 #:: 2 #:: 3 #:: LazyList.Empty
